Renamed original Main.py to test_main.py. Created a new Main.py that will be cleaned up
